Add functions to support notifying user program when screen is active
The user program is only allowed to draw on the screen when the screen
is "active". For performance reasons, we don't want to check this in
each drawing function. So it is up to the user program to use these
to know when it is OK to draw directly on the screen.