Better handling of linux console switching
Unsafe functions were being called from the SIGUSR1 handler. For example
it was possible to call malloc for allocating a new context in the middle
of another malloc.

It is now handled by using the GLib unix signal handler which calls
back as part of the main loop rather than directly from the low level
signal handler. There is also a property that can be used for notification
to the user program when we loose control of the console. It is up to
the user program to stop writing to the framebuffer at this point.