Refactor application classes into a single class
GrxLinuxConsoleApplication and GrxGtk3Application are now combined
into a single class. This required modification of the event system
to handle application events to indicate when an application is
activated or deactivated and when it requests to quit.

It also required moving the device manager instanciation into the
video driver init. There is now a global getter for the device manager.