Replace font handling with fontconfig and freetype2
There is no reason to redistribute fonts that are on the system in a
different format and do all of our own font handling when there are
libraries that already do this.

This removes quite a bit of functionality, but it should not be too hard
to add it back if needed.