Fix the linux fb driver
The driver was not working because it was closing ttyfd before it was
supposed to. Also rearranged the order of things to work like SDL 1.2.
This way, if we run as root, we don't have to have a vt already, or if
we run as a normal user, it uses the current vt.